Move time functions to time.py
All time related functions have been moved from functions.py to utils/time.py.
